factories for composer attachments
