+++ packages/stream_chat_flutter/lib/src/components/message_composer/message_composer_input_header.dart
@@ -101,7 +101,7 @@ class _DefaultStreamMessageComposerInputHeader extends StatelessWidget {
                 ),
               ),
             if (hasAttachments)
-              StreamMessageInputAttachmentList(
+              StreamMessageComposerAttachmentList(
                 attachments: nonOGAttachments,
                 onRemovePressed: _onAttachmentRemovePressed,
               ),
