+++ packages/stream_chat_flutter/lib/src/message_input/stream_message_composer_attachment.dart
@@ -1,0 +1,257 @@
+part of 'stream_message_composer_attachment_list.dart';
+
+class StreamMessageComposerAttachment extends StatelessWidget {
+  StreamMessageComposerAttachment({
+    super.key,
+    required Attachment attachment,
+    ValueSetter<Attachment>? onRemovePressed,
+    StreamAudioPlaylistController? audioPlaylistController,
+  }) : props = StreamMessageComposerAttachmentProps(
+         attachment: attachment,
+         onRemovePressed: onRemovePressed,
+         audioPlaylistController: audioPlaylistController,
+       );
+
+  /// The properties for the message composer attachment.
+  final StreamMessageComposerAttachmentProps props;
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
+
+class StreamMessageComposerAttachmentProps {
+  const StreamMessageComposerAttachmentProps({
+    required this.attachment,
+    required this.onRemovePressed,
+    required this.audioPlaylistController,
+  });
+
+  final Attachment attachment;
+  final ValueSetter<Attachment>? onRemovePressed;
+  final StreamAudioPlaylistController? audioPlaylistController;
+}
+
+class DefaultMessageComposerAttachment extends StatelessWidget {
+  const DefaultMessageComposerAttachment({super.key, required this.props});
+
+  final StreamMessageComposerAttachmentProps props;
+
+  Attachment get attachment => props.attachment;
+  ValueSetter<Attachment>? get onRemovePressed => props.onRemovePressed;
+  StreamAudioPlaylistController? get audioPlaylistController => props.audioPlaylistController;
+
+  @override
+  Widget build(BuildContext context) {
+    if (attachment.type == AttachmentType.file) {
+      return SizedBox(
+        width: 268,
+        child: MessageComposerFileAttachment(
+          fileTypeIcon: StreamFileTypeIcon.fromMimeType(mimeType: attachment.file?.mediaType?.mimeType ?? ''),
+          title: attachment.title ?? context.translations.fileText,
+          subtitle: _FileAttachmentSubtitle(attachment: attachment),
+          onRemovePressed: onRemovePressed != null ? () => onRemovePressed!(attachment) : null,
+        ),
+      );
+    }
+
+    if (attachment.type == AttachmentType.audio || attachment.type == AttachmentType.voiceRecording) {
+      if (audioPlaylistController == null) {
+        return const SizedBox.shrink();
+      }
+
+      final hasTrack = audioPlaylistController!.value.tracks.any((it) => it.key == attachment);
+
+      if (!hasTrack) {
+        return const SizedBox.shrink();
+      }
+
+      final trackIndex = audioPlaylistController!.value.tracks.indexWhere((it) => it.key == attachment);
+
+      return SizedBox(
+        width: 268,
+        child: MessageInputVoiceRecordingAttachment(
+          attachment: attachment,
+          index: trackIndex,
+          controller: audioPlaylistController!,
+          onRemovePressed: onRemovePressed,
+        ),
+      );
+    }
+
+    return StreamMediaAttachmentBuilder(
+      attachment: attachment,
+      onRemovePressed: onRemovePressed,
+    );
+  }
+}
+
+class _FileAttachmentSubtitle extends StatelessWidget {
+  const _FileAttachmentSubtitle({
+    required this.attachment,
+  });
+
+  final Attachment attachment;
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = StreamChatTheme.of(context);
+    final size = attachment.file?.size ?? attachment.extraData['file_size'];
+    final textStyle = theme.textTheme.footnote.copyWith(
+      color: theme.colorTheme.textLowEmphasis,
+    );
+    return attachment.uploadState.when(
+      preparing: () => Text(fileSize(size), style: textStyle),
+      inProgress: (sent, total) => StreamUploadProgressIndicator(
+        uploaded: sent,
+        total: total,
+        showBackground: false,
+        textStyle: textStyle,
+        progressIndicatorColor: theme.colorTheme.accentPrimary,
+      ),
+      success: () => Text(fileSize(size), style: textStyle),
+      failed: (_) => Text(
+        context.translations.uploadErrorLabel,
+        style: textStyle,
+      ),
+    );
+  }
+}
+
+/// Widget used to display the list of voice recording type attachments added to
+/// the message input.
+class MessageInputVoiceRecordingAttachment extends StatelessWidget {
+  /// Creates a new MessageInputVoiceRecordingAttachments widget.
+  const MessageInputVoiceRecordingAttachment({
+    super.key,
+    required this.attachment,
+    required this.index,
+    required this.controller,
+    this.onRemovePressed,
+  });
+
+  /// Attachment to display.
+  final Attachment attachment;
+
+  /// Index of the track in the playlist.
+  final int index;
+
+  /// Controller to use to control the audio playback.
+  final StreamAudioPlaylistController controller;
+
+  /// Callback called when the remove button is pressed.
+  final ValueSetter<Attachment>? onRemovePressed;
+
+  @override
+  Widget build(BuildContext context) {
+    return ValueListenableBuilder(
+      valueListenable: controller,
+      builder: (context, state, _) {
+        final track = state.tracks.where((it) => it.key == attachment).first;
+
+        return StreamVoiceRecordingAttachment(
+          title: 'Voice Message',
+          showTitle: true,
+          track: track,
+          speed: state.speed,
+          onRemovePressed: switch (onRemovePressed) {
+            final callback? => () => callback(attachment),
+            _ => null,
+          },
+          onTrackPause: controller.pause,
+          onChangeSpeed: controller.setSpeed,
+          onTrackPlay: () async {
+            // Play the track directly if it is already loaded.
+            if (state.currentIndex == index) return controller.play();
+            // Otherwise, load the track first and then play it.
+            return controller.skipToItem(index);
+          },
+          // Only allow seeking if the current track is the one being
+          // interacted with.
+          onTrackSeekStart: (_) async {
+            if (state.currentIndex != index) return;
+            return controller.pause();
+          },
+          onTrackSeekEnd: (_) async {
+            if (state.currentIndex != index) return;
+            return controller.play();
+          },
+          onTrackSeekChanged: (progress) async {
+            if (state.currentIndex != index) return;
+
+            final duration = track.duration.inMicroseconds;
+            final seekPosition = (duration * progress).toInt();
+            final seekDuration = Duration(microseconds: seekPosition);
+
+            return controller.seek(seekDuration);
+          },
+        );
+      },
+    );
+  }
+}
+
+/// Widget used to display a media type attachment item.
+class StreamMediaAttachmentBuilder extends StatelessWidget {
+  /// Creates a new media attachment item.
+  const StreamMediaAttachmentBuilder({super.key, required this.attachment, this.onRemovePressed});
+
+  /// The media attachment to display.
+  final Attachment attachment;
+
+  /// Callback called when the remove button is pressed.
+  final ValueSetter<Attachment>? onRemovePressed;
+
+  @override
+  Widget build(BuildContext context) {
+    final durationSecs = attachment.extraData['duration'] as num?;
+    final videoDuration = durationSecs != null ? Duration(seconds: durationSecs.round()) : null;
+
+    final mediaBadge = attachment.type == AttachmentType.video
+        ? StreamMediaBadge(type: MediaBadgeType.video, duration: videoDuration)
+        : null;
+
+    return Container(
+      key: Key(attachment.id),
+      child: MessageComposerMediaFileAttachment(
+        mediaBadge: mediaBadge,
+        onRemovePressed: onRemovePressed != null ? () => onRemovePressed!(attachment) : null,
+        child: StreamMediaAttachmentThumbnail(
+          media: attachment,
+          width: double.infinity,
+          height: double.infinity,
+          fit: BoxFit.cover,
+        ),
+      ),
+    );
+  }
+}
+
+/// Material Button used for removing attachments.
+class RemoveAttachmentButton extends StatelessWidget {
+  /// Creates a new remove attachment button.
+  const RemoveAttachmentButton({super.key, this.onPressed});
+
+  /// Callback when the remove attachment button is pressed.
+  final VoidCallback? onPressed;
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = StreamChatTheme.of(context);
+    final colorTheme = theme.colorTheme;
+
+    return IconButton.filled(
+      onPressed: onPressed,
+      color: colorTheme.barsBg,
+      padding: EdgeInsets.zero,
+      icon: Icon(context.streamIcons.crossMedium),
+      style: IconButton.styleFrom(
+        minimumSize: const Size(24, 24),
+        tapTargetSize: MaterialTapTargetSize.shrinkWrap,
+        // ignore: deprecated_member_use
+        backgroundColor: colorTheme.textHighEmphasis.withOpacity(0.6),
+      ),
+    );
+  }
+}
