+++ packages/stream_chat_flutter/lib/stream_chat_flutter.dart
@@ -102,8 +102,8 @@ export 'src/message_input/audio_recorder/stream_audio_recorder.dart';
 export 'src/message_input/countdown_button.dart';
 export 'src/message_input/enums.dart';
 export 'src/message_input/quoted_message_widget.dart';
+export 'src/message_input/stream_message_composer_attachment_list.dart';
 export 'src/message_input/stream_message_input.dart';
-export 'src/message_input/stream_message_input_attachment_list.dart';
 export 'src/message_input/stream_message_send_button.dart';
 export 'src/message_input/stream_message_text_field.dart';
 export 'src/message_list_view/message_details.dart';
